+++ waspvii/VIInterpreter.i.h
@@ -169,8 +169,19 @@ bool VIInterpreter<S>::load_document(size_t node_index, const std::string& path)
 
     std::string directory_name = wasp::dir_name(Interpreter<S>::stream_name());
     if (directory_name == Interpreter<S>::stream_name()) directory_name=".";
-    auto document_relative_path  = directory_name+"/"+path;
-    if ( wasp::file_exists(document_relative_path) )
+    auto document_relative_path  = directory_name + "/" + path;
+    // if immediately adjacent path doesn't exist
+    // check the VERA's Init directory
+    if (!wasp::file_exists(document_relative_path))
+    {
+        std::string vera_dir = wasp::get_env("VERA_DIR");
+        if (!vera_dir.empty())
+        {
+            document_relative_path = vera_dir + "/Init/" + path;
+        }
+    }
+    
+    if (wasp::file_exists(document_relative_path))
     {
         auto * interp = new VIInterpreter<S>(err_msgs);
         interp->m_parent = this;