+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - support for more than one input and template file in waspdrive input 
  - support for VERA input section naming and validation
  - support for parsing VERA input list syntax `<n..m x i> 193*10` etc.
+ - support for parsing VERA include files located in $VERA_DIR/Init. Requires export of VERA_DIR environment variable such that $VERA_DIR/Init exists
 
 
 ### Fixed